Add system methods to system lib
