+++ Sources/TCPSocket.swift
@@ -8,35 +8,6 @@
 
 import Foundation
 
-#if os(Linux)
-    import Glibc
-    let socketConnect = Glibc.connect
-    let socketBind = Glibc.bind
-    let socketListen = Glibc.listen
-    let socketAccept = Glibc.accept
-    let socketSend = Glibc.send
-    let recvRecv = Glibc.recv
-    let shutdown = Glibc.shutdown
-    let socketClose = Glibc.close
-    let getpeername = Glibc.getpeername
-    let getsockname = Glibc.getsockname
-    #else
-import Darwin
-    let isLittleEndian = Int(OSHostByteOrder()) == OSLittleEndian
-    let htons  = isLittleEndian ? _OSSwapInt16 : { $0 }
-    let ntohs  = isLittleEndian ? _OSSwapInt16 : { $0 }
-    let socketConnect = Darwin.connect
-    let socketBind = Darwin.bind
-    let socketListen = Darwin.listen
-    let socketAccept = Darwin.accept
-    let socketSend = Darwin.send
-    let recvRecv = Darwin.recv
-    let shutdown = Darwin.shutdown
-    let socketClose = Darwin.close
-    let getpeername = Darwin.getpeername
-    let getsockname = Darwin.getsockname
-#endif
-
 /// Class wrapping around TCP/IPv6 socket
 public final class TCPSocket {
     /// The file descriptor number for socket
@@ -144,7 +115,7 @@ public final class TCPSocket {
         // bind the address and port on socket
         guard withUnsafePointer(to: &address, { pointer in
             return pointer.withMemoryRebound(to: sockaddr.self, capacity: Int(size)) { pointer in
-                return socketBind(fileDescriptor, pointer, size) >= 0
+                return SystemLibrary.bind(fileDescriptor, pointer, size) >= 0
             }
         }) else {
             throw OSError.lastIOError()
@@ -154,7 +125,7 @@ public final class TCPSocket {
     /// Listen incomming connections
     ///  - Parameter backlog: maximum backlog of incoming connections
     func listen(backlog: Int = Int(SOMAXCONN)) throws {
-        guard socketListen(fileDescriptor, Int32(backlog)) != -1 else {
+        guard SystemLibrary.listen(fileDescriptor, Int32(backlog)) != -1 else {
             throw OSError.lastIOError()
         }
     }
@@ -165,7 +136,7 @@ public final class TCPSocket {
         var size = socklen_t(MemoryLayout<sockaddr_in6>.size)
         let clientFileDescriptor = withUnsafeMutablePointer(to: &address) { pointer in
             return pointer.withMemoryRebound(to: sockaddr.self, capacity: Int(size)) { pointer in
-                return socketAccept(fileDescriptor, pointer, &size)
+                return SystemLibrary.accept(fileDescriptor, pointer, &size)
             }
         }
         guard clientFileDescriptor >= 0 else {
@@ -192,7 +163,7 @@ public final class TCPSocket {
         // connect to the host and port
         let connectResult = withUnsafePointer(to: &address) { pointer in
             return pointer.withMemoryRebound(to: sockaddr.self, capacity: Int(size)) { pointer in
-                return socketConnect(fileDescriptor, pointer, size)
+                return SystemLibrary.connect(fileDescriptor, pointer, size)
             }
         }
         guard connectResult >= 0 || errno == EINPROGRESS else {
@@ -206,7 +177,7 @@ public final class TCPSocket {
     @discardableResult
     func send(data: Data) throws -> Int {
         let bytesSent = data.withUnsafeBytes { pointer in
-            socketSend(fileDescriptor, pointer, data.count, Int32(0))
+            SystemLibrary.send(fileDescriptor, pointer, data.count, Int32(0))
         }
         guard bytesSent >= 0 else {
             throw OSError.lastIOError()
@@ -220,7 +191,7 @@ public final class TCPSocket {
     func recv(size: Int) throws -> Data {
         var bytes = Data(count: size)
         let bytesRead = bytes.withUnsafeMutableBytes { pointer in
-            return recvRecv(fileDescriptor, pointer, size, Int32(0))
+            return SystemLibrary.recv(fileDescriptor, pointer, size, Int32(0))
         }
         guard bytesRead >= 0 else {
             throw OSError.lastIOError()
@@ -230,8 +201,8 @@ public final class TCPSocket {
 
     /// Close the socket
     func close() {
-        _ = shutdown(fileDescriptor, Int32(SHUT_WR))
-        _ = socketClose(fileDescriptor)
+        _ = SystemLibrary.shutdown(fileDescriptor, Int32(SHUT_WR))
+        _ = SystemLibrary.close(fileDescriptor)
     }
 
     func getPeerName() throws -> (String, Int) {
@@ -269,7 +240,7 @@ public final class TCPSocket {
                             family: AF_INET,
                             addressLength: INET_ADDRSTRLEN
                         ),
-                        Int(ntohs(addressptr.pointee.sin_port))
+                        Int(SystemLibrary.ntohs(addressptr.pointee.sin_port))
                     )
                 }
             case AF_INET6:
@@ -283,7 +254,7 @@ public final class TCPSocket {
                             family: AF_INET6,
                             addressLength: INET6_ADDRSTRLEN
                         ),
-                        Int(ntohs(addressptr.pointee.sin6_port))
+                        Int(SystemLibrary.ntohs(addressptr.pointee.sin6_port))
                     )
                 }
             default:
